Pass MySQL user/password to csr1kv_install_all.sh
If MySQL is not set up with passwordless access to all
databases, then the MySQL commands run by scipts called
from csr1kv_install_all.sh need to use the username and
password arguments.

This takes the localrc settings for MySQL user and password
(which need to be set with values and not reference other
variables), and passes them down to csr1kv_install_all.sh.

A companion change is needed in csr1kv_install_all.sh and
associated scripts in the Neutron repo.