Setup for CSR1kv, before Neutron network setup
Moving the logic in local.sh to lib/neutron so that
the CSR1kv setup steps can be done before DevStack
creates the initial network (in particular, the router).

Also, changed the logic so that the plugin variable
passed to the csr1kv_install_all.sh is set to n1kv,
when Q_PLUGIN=cisco. Currently, the logic appears to
be reversed and the script exits, in the case where
the n1kv plugin is being used (e.g. when using the
lcoalrc.icehouse_n1kv_csr1kl as localrc).